Explore Interaction Terms function
This function is designed to aid in quick exploration of a possible interaction term. 
yourdata = a dataframe containing all the variables you wish to include in your exploration. Include only columns of target, interaction1, and interaction2 if you wish to explore the terms in isolation from the rest of the model you are building. 

This is version_1